# src/main/resources/swagger.yaml
openapi: 3.0.3
info:
  title: Bank Application - API Specification
  description: |-
    This is a sample Bank Application developed with Spring Boot based on the OpenAPI 3.0.3 specification.  You can find out more about
    Swagger at [https://swagger.io](https://swagger.io).

    Some useful links:
    - [The Bank Application repository](https://github.com/Azadap27/bank-application)
    - [The source API definition for the Bank Application](https://github.com/Azadap27/bank-application/blob/master/src/main/resources/swagger.yaml)

  termsOfService: http://swagger.io/terms/
  contact:
    email: apiteam@swagger.io
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.0
externalDocs:
  description: Find out more about Swagger
  url: http://swagger.io
tags:
  - name: account
    description: Operations on account
  - name: person
    description: Operation about person
  - name: transfer
    description: Operation about bank transfer transaction
servers:
  - url: http://localhost:8080/api/v1
paths:
  /accounts:
    post:
      tags:
        - account
      summary: Create a new account in the bank
      description: Create a new account in the bank.
      operationId: createAccount
      requestBody:
        description: Create a new account in the bank
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AccountRequest'
        required: true
      responses:
        '200':
          description: Account Created Successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountResponse'
        '400':
          description: Bad request
        '405':
          description: Method not allowed
  /accounts/{accountId}:
    get:
      tags:
        - account
      summary: Get account details
      description: Get account details including account balance.
      operationId: getAccount
      parameters:
        - name: accountId
          in: path
          description: ID of account to return
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Account details retrieved successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountResponse'
        '400':
          description: Invalid Account ID supplied
        '404':
          description: Account not found
    post:
      tags:
        - account
      summary: Add balance in account
      description: Add balance in account.
      operationId: addBalance
      parameters:
        - name: accountId
          in: path
          description: ID of account in which balance to add
          required: true
          schema:
            type: integer
            format: int64
        - name: amount
          in: query
          description: Amount that needs to be added to account balance
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Balance added to account successfully.
        '400':
          description: Invalid Account ID supplied
        '404':
          description: Account not found
  /accounts/person/{personId}:
    get:
      tags:
        - person
      summary: Retrieve person information
      description: Get person details.
      operationId: getPerson
      parameters:
        - name: personId
          in: path
          description: ID of person to return
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Person details retrieved successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PersonResponse'
        '400':
          description: Invalid Person ID supplied
        '404':
          description: Person not found
  /accounts/transfer:
    post:
      tags:
        - transfer
      summary: Transfer amount to another account
      description: Transfer amount from one account to another account.
      operationId: transferAmount
      requestBody:
        description: Transfer amount to another account
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TransferRequest'
        required: true
      responses:
        '200':
          description: Amount Transferred Successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransferResponse'
        '400':
          description: Bad request
        '405':
          description: Method not allowed
components:
  schemas:
    PersonRequest:
      type: object
      required:
        - firstName
        - lastName
        - email
        - password
        - phoneNumber
      properties:
        firstName:
          type: string
          example: John
        lastName:
          type: string
          example: James
        email:
          type: string
          example: john@email.com
        password:
          type: string
          example: '12345'
        phoneNumber:
          type: string
          example: '12345'
    AccountRequest:
      type: object
      required:
        - person
      properties:
        person:
          $ref: '#/components/schemas/PersonRequest'
        initialCredit:
          type: integer
          format: int64
          example: 500
    TransferRequest:
      type: object
      required:
        - senderAccount
        - receiverAccount
        - amount
      properties:
        senderAccount:
          type: integer
          format: int64
        receiverAccount:
          type: integer
          format: int64
        amount:
          type: integer
          format: int64
    PersonResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 10
        firstName:
          type: string
          example: John
        lastName:
          type: string
          example: James
        email:
          type: string
          example: john@email.com
        phoneNumber:
          type: string
          example: '12345'
    AccountResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 5
        person:
          $ref: '#/components/schemas/PersonResponse'
        balance:
          type: integer
          format: int64
          example: 1000
    TransferResponse:
      type: object
      properties:
        transactionId:
          type: integer
          format: int64
        senderAccount:
          type: integer
          format: int64
        receiverAccount:
          type: integer
          format: int64
        amount:
          type: integer
          format: int64